refactor: move signal handlers and TTY check into cleanup.ts
Moves setupSignalHandlers(), setupTtyCheck(), gracefulShutdown(), and
isShuttingDown from gemini.tsx into utils/cleanup.ts to keep gemini.tsx
focused on app initialization. Cleanup-related logic now lives in one
place.
