+++ packages/cli/src/gemini.tsx
@@ -32,6 +32,8 @@ import {
   registerSyncCleanup,
   runExitCleanup,
   registerTelemetryConfig,
+  setupSignalHandlers,
+  setupTtyCheck,
 } from './utils/cleanup.js';
 import {
   cleanupToolOutputFiles,
@@ -179,90 +181,6 @@ ${reason.stack}`
       appEvents.emit(AppEvent.OpenDebugConsole);
     }
   });
-}
-
-/**
- * Sets up signal handlers to ensure graceful shutdown and prevent orphaned
- * processes from consuming 100% CPU when the terminal is closed.
- *
- * Handles:
- * - SIGHUP: Terminal hangup (terminal window closed)
- * - SIGTERM: Termination request
- * - SIGINT: Interrupt (Ctrl+C) - as fallback, Ink handles this in interactive mode
- *
- * @see https://github.com/google-gemini/gemini-cli/issues/15874
- */
-let isShuttingDown = false;
-
-const gracefulShutdown = async (reason: string) => {
-  // Prevent multiple concurrent shutdown attempts
-  if (isShuttingDown) {
-    return;
-  }
-  isShuttingDown = true;
-
-  debugLogger.debug(`Shutting down gracefully due to ${reason}...`);
-  try {
-    await runExitCleanup();
-  } catch (err) {
-    debugLogger.debug(`Error during cleanup for ${reason}:`, err);
-  }
-  process.exit(ExitCodes.SUCCESS);
-};
-
-export function setupSignalHandlers() {
-  // SIGHUP is sent when terminal is closed - critical for preventing orphans
-  process.on('SIGHUP', () => gracefulShutdown('SIGHUP'));
-  process.on('SIGTERM', () => gracefulShutdown('SIGTERM'));
-  // SIGINT is handled by Ink in interactive mode, but we need it for non-interactive
-  process.on('SIGINT', () => gracefulShutdown('SIGINT'));
-}
-
-/**
- * Sets up periodic TTY check to detect lost controlling terminal.
- * This is a defense-in-depth measure for cases where SIGHUP is not received
- * or not processed correctly (e.g., certain terminal emulators, tmux detach).
- *
- * @returns Cleanup function to stop the TTY check interval
- */
-export function setupTtyCheck(): () => void {
-  let intervalId: ReturnType<typeof setInterval> | null = null;
-  let isCheckingTty = false;
-
-  intervalId = setInterval(async () => {
-    // Prevent concurrent TTY checks
-    if (isCheckingTty || isShuttingDown) {
-      return;
-    }
-
-    // Skip check in sandbox mode or if explicitly running non-interactively
-    if (process.env['SANDBOX'] || process.env['GEMINI_NON_INTERACTIVE']) {
-      return;
-    }
-
-    // Check if we've lost both stdin and stdout TTY
-    if (!process.stdin.isTTY && !process.stdout.isTTY) {
-      isCheckingTty = true;
-
-      // Clear interval BEFORE starting cleanup to prevent race condition
-      if (intervalId) {
-        clearInterval(intervalId);
-        intervalId = null;
-      }
-
-      await gracefulShutdown('TTY loss');
-    }
-  }, 5000);
-
-  // Don't keep the process alive just for this interval
-  intervalId.unref();
-
-  return () => {
-    if (intervalId) {
-      clearInterval(intervalId);
-      intervalId = null;
-    }
-  };
 }
 
 export async function startInteractiveUI(
